homepage UX fixes
