+++ app/assets/stylesheets/layout.scss
@@ -61,7 +61,7 @@ button:hover, button:focus {
 /* LoggedOutMessage */
 
 h1.greeting {
-  margin: 2%;
+  margin: 2% 1.5%;
   font-size: 20px;
   font-family: 'Source Sans Pro', sans-serif;
   font-weight: bold;
