+++ app/javascript/containers/MediaIndexContainer.js
@@ -1,25 +1,18 @@
 import React, { Component } from 'react'
 import { browserHistory } from 'react-router'
 
-import MediaPreview from '../components/MediaPreview'
-import NewMediaButton from '../components/NewMediaButton'
-import PageButton from '../components/PageButton'
 import FlashNotice from '../components/FlashNotice'
+import MediaIndexTile from '../components/MediaIndexTile'
+import AddMediaPrompt from '../components/AddMediaPrompt'
 
 class MediaIndexContainer extends Component {
   constructor(props) {
     super(props)
     this.state = {
-      userId: null,
-      firstName: null,
-      lastName: null,
       media: [],
-      slicePoint1: 0,
-      slicePoint2: 6,
-      slicePoint3: 12
+      loaded: false
     }
 
-    this.pageFlip = this.pageFlip.bind(this)
     this.clearFlash = this.clearFlash.bind(this)
   }
 
@@ -39,30 +32,13 @@ class MediaIndexContainer extends Component {
     })
     .then(body => {
       this.setState({
-        userId: body.user.id,
-        firstName: body.user.first_name,
-        lastName: body.user.last_name,
-        media: body.user.media
+        media: body.user.media,
+        loaded: true
       })
     })
     .catch(error => {
       console.error(`Error in fetch: ${error.message}`)
     });
-  }
-
-  pageFlip(value) {
-    event.preventDefault()
-    let flip = (value === "right" ? 12 : -12)
-
-    let currentSlice1 = this.state.slicePoint1
-    let currentSlice2 = this.state.slicePoint2
-    let currentSlice3 = this.state.slicePoint3
-
-    this.setState({
-      slicePoint1: currentSlice1 + flip,
-      slicePoint2: currentSlice2 + flip,
-      slicePoint3: currentSlice3 + flip
-    })
   }
 
   clearFlash() {
@@ -73,15 +49,7 @@ class MediaIndexContainer extends Component {
   }
 
   render() {
-    let media = this.state.media
-
-    let flashNotice
-    let topPreviewTiles
-    let bottomPreviewTiles
-    let addMedia
-    let backButton
-    let nextButton
-
+    let flashNotice, displayComponent
     if (this.props.location.state) {
       flashNotice =
         <FlashNotice
@@ -90,64 +58,20 @@ class MediaIndexContainer extends Component {
         />
     }
 
-    if (media.length !== 0) {
-      let key = 0
-      let mediaPreviewTiles = media.map((media_object) => {
-        let type = `${media_object.director ? "movie" : "show"}`
-        key++
-
-        return(
-          <MediaPreview
-            key={key}
-            id={media_object.id}
-            name={media_object.name}
-            poster={media_object.poster.url}
-            type={type}
-          />
-        )
-      })
-
-      topPreviewTiles = mediaPreviewTiles.slice(this.state.slicePoint1, this.state.slicePoint2)
-      bottomPreviewTiles = mediaPreviewTiles.slice(this.state.slicePoint2, this.state.slicePoint3)
-
-      if (this.state.slicePoint1 !== 0 ) {
-        backButton =
-          <PageButton
-            direction="left"
-            pageFlip={ this.pageFlip }
-          />
-      }
-
-      if (this.state.slicePoint3 < media.length) {
-        nextButton =
-          <PageButton
-            direction="right"
-            pageFlip={ this.pageFlip }
-          />
-      }
+    if (this.state.media.length !== 0) {
+      displayComponent =
+        <MediaIndexTile
+          media={this.state.media}
+        />
     }
-
-    addMedia =
-      <NewMediaButton
-        selectedClass="add-media"
-      />
+    else if (this.state.media.length === 0 && this.state.loaded === true){
+      displayComponent = <AddMediaPrompt />
+    }
 
     return(
       <div>
         {flashNotice}
-        <div className="homepage">
-          <div className="main-col small-12 large-12 columns">
-            <div>{ topPreviewTiles }</div>
-            <div>{ bottomPreviewTiles }</div>
-          </div>
-          <div className="homepage-options">
-            {addMedia}
-            <div className="pagination">
-              {backButton}
-              {nextButton}
-            </div>
-          </div>
-        </div>
+        {displayComponent}
       </div>
     )
   }
